cpu/xmc1000: de-inline usic_init() and reorganize headers
