+++ cpu/xmc1000/include/periph/usic.h
@@ -11,7 +11,7 @@
  * @{
  *
  * @file
- * @brief           Code shared by peripheral drivers using the USIC
+ * @brief           Definitions for code shared by peripheral drivers using the USIC
  *
  * @author          Sebastian Sontberg <sebastian@sontberg.de>
  *
@@ -21,9 +21,7 @@
 #ifndef USIC_H_
 #define USIC_H_
 
-#include "cpu.h"
-#include "periph/gating.h"
-#include "periph/usic_defs.h"
+#include "periph_cpu.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -34,63 +32,143 @@ extern "C" {
  * @ingroup cpu_xmc1000
  * @{
  */
+#define USIC_CTQIN_PDIV  (0)    /**< f<sub>CTQIN</sub> is provided by f<sub>PDIV</sub> */
+#define USIC_CTQIN_PPP   (1)    /**< f<sub>CTQIN</sub> is provided by f<sub>PPP</sub>  */
+#define USIC_CTQIN_SCLK  (2)    /**< f<sub>CTQIN</sub> is provided by f<sub>SCLK</sub> */
+#define USIC_CTQIN_MCLK  (3)    /**< f<sub>CTQIN</sub> is provided by f<sub>MCLK</sub> */
+
+#define USIC_MODE_ASC    (0)    /**< select ASC/UART mode */
+#define USIC_MODE_SSC    (1)    /**< select SSC/SPI mode  */
+
+/**
+ * @brief Struct to hold the configuration values necessary to
+ *        initialize a USIC channel
+ * @{
+ */
+typedef struct {
+    uint32_t ccr;       /**< channel control register           */
+    uint32_t sctr;      /**< shift control register             */
+    uint32_t tcsr;      /**< transmit control/status register   */
+    uint32_t pcr;       /**< protocol control register          */
+    uint32_t inpr;      /**< interrupt node pointer register    */
+    uint8_t dx2_dsel;   /**< data selection for input stage DX2 */
+} usic_mode_t;
+/** @} */
+
+/**
+ * @brief   USIC BRG (baud rate generator) register configuration
+ * @{
+ */
+typedef struct {
+    union {
+        struct __attribute__((packed)) {
+            unsigned clksel : 2;
+            unsigned __reserved1 : 1;
+            unsigned tmen : 1;
+            unsigned pppen : 1;
+            unsigned __reserved2 : 1;
+            unsigned ctqsel : 2;
+            unsigned pctq : 2;
+            unsigned dctq : 5;
+            unsigned __reserved3 : 1;
+            unsigned pdiv : 10;
+            unsigned __reserved4 : 2;
+            unsigned sclkosel : 1;
+            unsigned mclkcfg : 1;
+            unsigned sclkcfg : 2;
+        } field;
+        uint32_t reg;
+    };
+} usic_brg_t;
+/** @} */
+
+/**
+ * @brief   USIC FDR (fractional divider) register configuration
+ * @{
+ */
+typedef struct {
+    union {
+        struct __attribute__((packed)) {
+            unsigned step : 10;
+            unsigned __reserved1 : 4;
+            unsigned dm : 2;
+            unsigned result : 10;
+            unsigned __reserved2 : 6;
+        } field;
+        uint32_t reg;
+    };
+} usic_fdr_t;
+/** @} */
+
+/**
+ * @brief   USIC channel FIFO configuration
+ * @{
+ */
+typedef struct __attribute__((packed)) {
+    uint8_t rx_dptr;                /**< the index of the first entry of the receive FIFO    */
+    uint8_t rx_size;                /**< the size of the receive FIFO                        */
+    uint8_t tx_dptr;                /**< the index of the first entry of the transmit FIFO   */
+    uint8_t tx_size;                /**< the size of the transmit FIFO                       */
+} usic_fifo_t;
+/** @} */
+
+/**
+ * @brief   UART device configuration
+ * @{
+ */
+typedef struct __attribute__((packed)) {
+    USIC_CH_TypeDef *usic;          /**< pointer to USIC channel                             */
+    const usic_mode_t *mode;        /**< pointer to the control mode for this channel        */
+    const usic_fifo_t fifo;         /**< configuration of the FIFOs used by this channel     */
+    const gpio_alt_t tx_pin;        /**< the primary data output pin (MOSI/TX)               */
+    const gpio_alt_t rx_pin;        /**< the primary data input pin (MISO/RX)                */
+} uart_instance_t;
+
+#define UART_NUMOF         ((sizeof(uart_instance) / sizeof(uart_instance[0])))
+/** @} */
+
+/**
+ * @brief   SPI device configuration
+ * @{
+ */
+typedef struct __attribute__((packed)) {
+    USIC_CH_TypeDef *usic;          /**< pointer to USIC channel                             */
+    const usic_mode_t *mode;        /**< pointer to the control mode for this USIC           */
+    const usic_fifo_t fifo;         /**< configuration of the FIFOs used by this channel     */
+    const gpio_alt_t mosi_pin;      /**< MOSI pin                                            */
+    const gpio_alt_t miso_pin;      /**< MISO pin                                            */
+    const gpio_alt_t sclk_pin;      /**< master shift clock pin (SCLK)                       */
+    const gpio_alt_t msls_pin;      /**< master-slave-select pin (slave select)              */
+} spi_instance_t;
+/** @} */
+
+/**
+ * @brief   USIC channel configuration
+ * @{
+ */
+typedef struct __attribute__((packed)) {
+    USIC_CH_TypeDef *usic;          /**< pointer to USIC channel                             */
+    const usic_mode_t *mode;        /**< pointer to the control mode for this USIC           */
+    const usic_fifo_t fifo;         /**< configuration of the FIFOs used by this channel     */
+    const gpio_alt_t tx_pin;        /**< the primary data output pin (MOSI/TX)               */
+    const gpio_alt_t rx_pin;        /**< the primary data input pin (MISO/RX)                */
+    const gpio_alt_t sclk_pin;      /**< supplementary pin for master/slave SSC/SPI (clock)  */
+    const gpio_alt_t msls_pin;      /**< supplementary pin for master SSC/SPI (slave select) */
+} usic_channel_t;
+/** @} */
+
+/** @} */
 
 /**
  * @brief   USIC channel initialization function
  * @{
  */
-static inline void usic_init(const usic_channel_t *usic_ch,
-                             const usic_brg_t brg,
-                             const usic_fdr_t fdr)
-{
-    USIC_CH_TypeDef *usic = usic_ch->usic;
-
-    const usic_mode_t *controls = usic_ch->mode;
-
-    /* initialize input pin */
-    gpio_init(usic_ch->rx_pin & 0xff, GPIO_DIR_IN, GPIO_NOPULL);
-
-    /* clear gating, enable clock */
-    GATING_CLEAR(USIC0);
-
-    /* enable the USIC module */
-    usic->KSCFG =
-        (1 << USIC_CH_KSCFG_MODEN_Pos) |
-        (1 << USIC_CH_KSCFG_BPMODEN_Pos);
-
-    /* clear channel control register (reset mode bit) */
-    usic->CCR = 0;
-
-    /* configure fractional divider */
-    usic->FDR = fdr.reg;
-
-    /* configure baud rate generator */
-    usic->BRG = brg.reg;
-
-    /* configure input stages DX0 & DX2 */
-    usic->DX0CR = (usic_ch->rx_pin >> 8) << USIC_CH_DX0CR_DSEL_Pos;
-    usic->DX2CR =  controls->dx2_dsel << USIC_CH_DX2CR_DSEL_Pos;
-
-    /* interrupt node point register */
-    usic->INPR = controls->inpr;
-    /* shift control register */
-    usic->SCTR = controls->sctr;
-    /* transmit control/status register */
-    usic->TCSR = controls->tcsr;
-    /* protocol control register */
-    usic->PCR  = controls->pcr;
-    /* channel control register */
-    usic->CCR  = controls->ccr;
-
-    /* ready to activate output pins */
-    gpio_init(usic_ch->tx_pin & 0xff, GPIO_DIR_OUT | (usic_ch->tx_pin >> 8), GPIO_NOPULL);
-}
-/** @} */
-
-/** @} */
+void usic_init(const usic_channel_t *usic_ch,
+               const usic_brg_t brg,
+               const usic_fdr_t fdr);
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* USIC_H_ */
+#endif /* USIC_DEFS_H_ */
